Created Locationtracker interface.
